+++ jmx/jmx-api/src/main/java/uk/gov/justice/services/jmx/api/command/ShutterCommand.java
@@ -1,6 +1,6 @@
 package uk.gov.justice.services.jmx.api.command;
 
-public class ShutterCommand extends BaseSystemCommand {
+public class ShutterCommand extends BaseSystemCommand implements ApplicationShutteringCommand {
 
     public static final String SHUTTER = "SHUTTER";
     private static final String DESCRIPTION = "Shutters the application to allow for maintenance.";