+++ jmx/jmx-command-handling/src/test/java/uk/gov/justice/services/jmx/bootstrap/blacklist/BlacklistedCommandsFilterTest.java
@@ -3,14 +3,14 @@ package uk.gov.justice.services.jmx.bootstrap.blacklist;
 import static com.google.common.collect.Sets.newHashSet;
 import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.assertThat;
-import static uk.gov.justice.services.jmx.api.command.CatchupCommand.CATCHUP;
+import static uk.gov.justice.services.jmx.api.command.EventCatchupCommand.CATCHUP;
 import static uk.gov.justice.services.jmx.api.command.IndexerCatchupCommand.INDEXER_CATCHUP;
 import static uk.gov.justice.services.jmx.api.command.PingCommand.PING;
 import static uk.gov.justice.services.jmx.api.command.RebuildCommand.REBUILD;
 import static uk.gov.justice.services.jmx.api.command.ShutterCommand.SHUTTER;
 import static uk.gov.justice.services.jmx.api.command.UnshutterCommand.UNSHUTTER;
 
-import uk.gov.justice.services.jmx.api.command.CatchupCommand;
+import uk.gov.justice.services.jmx.api.command.EventCatchupCommand;
 import uk.gov.justice.services.jmx.api.command.IndexerCatchupCommand;
 import uk.gov.justice.services.jmx.api.command.PingCommand;
 import uk.gov.justice.services.jmx.api.command.RebuildCommand;
@@ -37,7 +37,7 @@ public class BlacklistedCommandsFilterTest {
         final Set<SystemCommand> blacklistedCommands = newHashSet(
                 new ShutterCommand(),
                 new UnshutterCommand(),
-                new CatchupCommand(),
+                new EventCatchupCommand(),
                 new PingCommand(),
                 new RebuildCommand(),
                 new IndexerCatchupCommand()